🎨 댓글 날짜 포맷 함수 utils/commentDateUtils.ts로 분리
