+++ client/src/components/CommentSection.tsx
@@ -13,6 +13,7 @@ import {
 import { Edit, Delete, ThumbUp, ThumbDown } from "@mui/icons-material";
 import { useAuthStore } from "../context/authStore";
 import getAnimalIcon from "../utils/getAnimalIcon";
+import { formatCommentDate } from "../utils/commentDateUtils";
 
 
 interface Comment {
@@ -232,21 +233,6 @@ const CommentSection = ({ postId }: CommentSectionProps) => {
       console.error("싫어요 취소 실패:", error);
     }
   };  
-
-  const formatCommentDate = (createdAt: string): string => {
-    const createdDate = new Date(createdAt);
-    const now = new Date();
-  
-    const isToday = createdDate.toDateString() === now.toDateString();
-  
-    return isToday
-      ? createdDate.toLocaleTimeString('ko-KR', {
-          hour: '2-digit',
-          minute: '2-digit',
-          hour12: false,
-        })
-      : `${createdDate.getFullYear()}.${String(createdDate.getMonth() + 1).padStart(2, '0')}.${String(createdDate.getDate()).padStart(2, '0')}`;
-  };
   
   if (loading) return <CircularProgress sx={{ display: "block", margin: "auto", mt: 3 }} />;
